requirements.txt
pytest
snowflake-snowpark-python
pandas
requests
snowflake-connector-python
pydantic